pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "irpropycapture"
version = "0.1.0"
description = "Python thermal camera viewer for InfiRay P2 Pro"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "IrProPyCapture contributors" },
]
dependencies = [
    "PySide6>=6.11",
    "numpy>=2.4",
    "opencv-python>=4.13",
    "pyusb>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
]

[project.scripts]
irpropycapture = "irpropycapture.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["irpropycapture*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
